int64_t
does it work?

+++ source/Personality.h
@@ -96,7 +96,7 @@ private:
 private:
 	bool isDefined = false;
 
-	int flags;
+	int64_t flags;
 	double confusionMultiplier;
 	double aimMultiplier;
 	Point confusion;
